evolution/requirements.txt
# Pinned versions — update periodically with: pip install --upgrade <pkg>
# Deus Evolution Loop dependencies
# Core (always required)
google-genai==1.14.0       # Gemini API (same as memory_indexer.py)
sqlite-vec==0.1.6          # Vector search in SQLite

# MCP server (optional — for external agent integration)
mcp==1.23.0

# DSPy optimizer (optional — install when running `evolve optimize`)
# dspy-ai>=2.5.0
# google-generativeai>=0.8.0   # DSPy's Gemini backend uses the older SDK
